+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+
 ## [0.16.0] - 2023-01-25
 ### Added
 - Added Clear Button to History and WebSockets Tab (Issue 411).
@@ -163,6 +166,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2018-12-03
 First alpha release.
 
+[Unreleased]: https://github.com/zaproxy/zap-hud/compare/v0.16.0...HEAD
 [0.16.0]: https://github.com/zaproxy/zap-hud/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/zaproxy/zap-hud/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/zaproxy/zap-hud/compare/v0.13.0...v0.14.0