+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [0.17.0] - 2023-07-11
+### Changed
+- Update minimum ZAP version to 2.13.0.
+
 ## [0.16.0] - 2023-01-25
 ### Added
 - Added Clear Button to History and WebSockets Tab (Issue 411).
@@ -166,7 +170,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2018-12-03
 First alpha release.
 
-[Unreleased]: https://github.com/zaproxy/zap-hud/compare/v0.16.0...HEAD
+[Unreleased]: https://github.com/zaproxy/zap-hud/compare/v0.17.0...HEAD
+[0.17.0]: https://github.com/zaproxy/zap-hud/compare/v0.16.0...v0.17.0
 [0.16.0]: https://github.com/zaproxy/zap-hud/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/zaproxy/zap-hud/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/zaproxy/zap-hud/compare/v0.13.0...v0.14.0